Pass back all callbacks to the active WrenHost
